Add angular and flot dependencies to bower.json
This allows tools like main-bower-files which use on bower.json metadata to insert and order your dependencies into your page.  Without this it was trying to add angular-flot.js before angular.js in the page, resulting in an error on the page.

I left the version requirements loose with a "*".  They could be restricted if desired.